requirement.txt
automata-lib==3.1.0.post1
autopep8==1.4.4
cycler==0.10.0
decorator==4.4.0
kiwisolver==1.1.0
networkx==2.3
numpy==1.17.2
pkg-resources==0.0.0
pycodestyle==2.5.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
termcolor==1.1.0
